pyproject.toml
[project]
name = "main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.8.1",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "colorama==0.4.6",
    "django==5.2.1",
    "django-cors-headers==4.7.0",
    "django-environ==0.12.0",
    "djangorestframework==3.16.0",
    "gunicorn==23.0.0",
    "idna==3.10",
    "imagekitio @ git+https://github.com/imagekit-developer/imagekit-python.git#egg=imagekit-python --no-deps",
    "packaging==25.0",
    "psycopg2-binary==2.9.10",
    "pycryptodome==3.23.0",
    "python-magic-bin==0.4.14",
    "requests==2.32.3",
    "requests-toolbelt>=0.10.1",
    "sqlparse==0.5.3",
    "tqdm==4.67.1",
    "tzdata==2025.2",
    "urllib3>=1.26.20",
    "vercel_blob @ git+https://github.com/SuryaSekhar14/vercel_blob.git#egg=vercel_blob --no-deps",
    "whitenoise==6.9.0",
]
